+++ Library/OcMainLib/OpenCoreKernel.c
@@ -39,6 +39,9 @@ STATIC BOOLEAN             mUse32BitKernel;
 STATIC CACHELESS_CONTEXT   mOcCachelessContext;
 STATIC BOOLEAN             mOcCachelessInProgress;
 
+STATIC EFI_FILE_PROTOCOL   *mCustomKernelDirectory;
+STATIC BOOLEAN             mCustomKernelDirectoryInProgress;
+
 STATIC
 VOID
 OcKernelConfigureCapabilities (
@@ -1078,6 +1081,13 @@ OcKernelFileOpen (
   UINT32             NumReservedKexts;
   UINT32             LinkedExpansion;
   UINT32             ReservedFullSize;
+  CHAR16             *NewFileName;
+  EFI_FILE_PROTOCOL  *EspNewHandle;
+
+  if (mCustomKernelDirectoryInProgress) {
+    DEBUG ((DEBUG_INFO, "OC: Skipping OpenFile hooking on ESP Kernels directory\n"));
+    return SafeFileOpen (This, NewHandle, FileName, OpenMode, Attributes);
+  }
 
   //
   // Prevent access to cache files depending on maximum cache type allowed.
@@ -1136,6 +1146,20 @@ OcKernelFileOpen (
     && Status == EFI_NOT_FOUND
     && OpenMode == EFI_FILE_MODE_READ
     && (StrStr (FileName, L"\\kernelcache") != NULL)) {
+    //
+    // Change the target to the custom one if requested CustomKernel.
+    //
+    if (mCustomKernelDirectory != NULL) {
+      DEBUG ((DEBUG_INFO, "OC: Redirecting %s to the custom one on ESP\n", FileName));
+      NewFileName = OcStrrChr (FileName, L'\\');
+      if (NewFileName == NULL) {
+        NewFileName = FileName;
+      }
+      DEBUG ((DEBUG_INFO, "OC: Filename after redirection: %s\n", NewFileName));
+
+      This     = mCustomKernelDirectory;
+      FileName = NewFileName;
+    }
 
     DEBUG ((DEBUG_INFO, "OC: Trying kernelcache fuzzy matching on %s\n", FileName));
 
@@ -1171,6 +1195,29 @@ OcKernelFileOpen (
     && StrCmp (FileName, L"System\\Library\\Kernels\\kernel") != 0
     && OcStriStr (FileName, L".kext\\") == NULL
     && OcStriStr (FileName, L".im4m") == NULL) {
+    //
+    // Change the target to the custom one if requested CustomKernel.
+    //
+    if (mCustomKernelDirectory != NULL) {
+      DEBUG ((DEBUG_INFO, "OC: Redirecting %s to the custom one on ESP\n", FileName));
+      NewFileName = OcStrrChr (FileName, L'\\');
+      if (NewFileName == NULL) {
+        NewFileName = FileName;
+      }
+
+      DEBUG ((DEBUG_INFO, "OC: Filename after redirection: %s\n", NewFileName));
+
+      mCustomKernelDirectoryInProgress = TRUE;
+      Status = SafeFileOpen (mCustomKernelDirectory, &EspNewHandle, NewFileName, OpenMode, Attributes);
+      mCustomKernelDirectoryInProgress = FALSE;
+      if (!EFI_ERROR (Status)) {
+        (*NewHandle)->Close (*NewHandle);
+
+        This = mCustomKernelDirectory;
+        *NewHandle = EspNewHandle;
+        FileName = NewFileName;
+      }
+    }
 
     //
     // Kernel loading for fuzzy kernelcache is performed earlier.
@@ -1191,7 +1238,7 @@ OcKernelFileOpen (
         );
 
       if (Status == EFI_NOT_FOUND) {
-        (*NewHandle)->Close(*NewHandle);
+        (*NewHandle)->Close (*NewHandle);
         *NewHandle = NULL;
 
         return Status;
@@ -1210,7 +1257,7 @@ OcKernelFileOpen (
         DEBUG ((DEBUG_INFO, "OC: Blocking prelinked due to ForceKernelCache=%s: %a\n", FileName, ForceCacheType));
 
         FreePool (Kernel);
-        (*NewHandle)->Close(*NewHandle);
+        (*NewHandle)->Close (*NewHandle);
         *NewHandle = NULL;
 
         return EFI_NOT_FOUND;
@@ -1248,7 +1295,7 @@ OcKernelFileOpen (
         ZeroMem (&ModificationTime, sizeof (ModificationTime));
       }
 
-      (*NewHandle)->Close(*NewHandle);
+      (*NewHandle)->Close (*NewHandle);
 
       //
       // Virtualise newly created kernel.
@@ -1280,7 +1327,7 @@ OcKernelFileOpen (
     //
     if (MaxCacheTypeAllowed == CacheTypeCacheless) {
       DEBUG ((DEBUG_INFO, "OC: Blocking mkext due to ForceKernelCache=%s: %a\n", FileName, ForceCacheType));
-      (*NewHandle)->Close(*NewHandle);
+      (*NewHandle)->Close (*NewHandle);
       *NewHandle = NULL;
 
       return EFI_NOT_FOUND;
@@ -1337,7 +1384,7 @@ OcKernelFileOpen (
           ZeroMem (&ModificationTime, sizeof (ModificationTime));
         }
 
-        (*NewHandle)->Close(*NewHandle);
+        (*NewHandle)->Close (*NewHandle);
 
         //
         // Virtualise newly created mkext.
@@ -1441,16 +1488,44 @@ OcLoadKernelSupport (
   IN OC_CPU_INFO         *CpuInfo
   )
 {
-  EFI_STATUS  Status;
+  EFI_STATUS         Status;
+  EFI_FILE_PROTOCOL  *Root;
 
   Status = EnableVirtualFs (gBS, OcKernelFileOpen);
 
   if (!EFI_ERROR (Status)) {
-    mOcStorage              = Storage;
-    mOcConfiguration        = Config;
-    mOcCpuInfo              = CpuInfo;
-    mOcDarwinVersion        = 0;
-    mOcCachelessInProgress  = FALSE;
+    mOcStorage                        = Storage;
+    mOcConfiguration                  = Config;
+    mOcCpuInfo                        = CpuInfo;
+    mOcDarwinVersion                  = 0;
+    mOcCachelessInProgress            = FALSE;
+    mCustomKernelDirectoryInProgress  = FALSE;
+    //
+    // Open customised Kernels if needed.
+    //
+    mCustomKernelDirectory = NULL;
+    if (mOcConfiguration->Kernel.Scheme.CustomKernel) {
+      Status = FindWritableOcFileSystem (&Root);
+      if (!EFI_ERROR (Status)) {
+        //
+        // Open Kernels directory.
+        //
+        Status = Root->Open (
+          Root,
+          &mCustomKernelDirectory,
+          L"Kernels",
+          EFI_FILE_MODE_READ,
+          EFI_FILE_DIRECTORY
+          );
+        if (EFI_ERROR (Status)) {
+          DEBUG ((DEBUG_INFO, "OC: Unable to open Kernels folder for custom kernel - %r, falling back to normal one\n", Status));
+          mCustomKernelDirectory = NULL;
+        }
+      } else {
+        DEBUG ((DEBUG_INFO, "OC: Unable to find root writable filesystem for custom kernel - %r, falling back to normal one\n", Status));
+      }
+    }
+
     OcImageLoaderRegisterConfigure (OcKernelConfigureCapabilities);
   } else {
     DEBUG ((DEBUG_ERROR, "OC: Failed to enable vfs - %r\n", Status));