+++ Changelog.md
@@ -12,6 +12,10 @@ OpenCore Changelog
 - Added explicit warn and safe fallback to builtin picker on failure to match the above
 - Added VSCode source level IDE debug config example to debug docs
 - Added other minor debug docs updates
+- Fixed incorrect timeout of built-in picker on IA32
+- Added support for custom kernels on ESP partition
+- Fixed DEBUG ASSERT on pressing change entry keys with single boot entry in OpenCanopy
+- Added recommended `Apple12` and `Windows11` flavours
 
 #### v0.7.0
 - Fixed NVRAM reset on firmware with write-protected `BootOptionSupport`